requirements.txt
tensorflow==2.6.0
keras==2.6.0        
pandas==1.3.3         
numpy==1.19.5
tqdm==4.62.3
uuid==1.30
pysftp==0.2.9
pillow==8.4.0
matplotlib==3.5.0
scikit-learn==0.24.2